requirements.txt
torch==1.8.1
torchvision==0.9.1
Pillow==8.1.0
jupyterlab==1.0.2
